Squashed migrations and created seed_roles.py
