--- alembic/versions/0845b2f0f401_add_request_task_order_relationship.py
@@ -1,30 +1,0 @@
-"""add request -> task order relationship
-
-Revision ID: 0845b2f0f401
-Revises: 875e4b8a05fc
-Create Date: 2018-08-22 09:58:43.770718
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '0845b2f0f401'
-down_revision = '875e4b8a05fc'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('requests', sa.Column('task_order_id', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'requests', 'task_order', ['task_order_id'], ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'requests', type_='foreignkey')
-    op.drop_column('requests', 'task_order_id')
-    # ### end Alembic commands ###
