--- alembic/versions/14cd800904bc_add_task_order_association_to_.py
@@ -1,30 +1,0 @@
-"""add task order association to attachments
-
-Revision ID: 14cd800904bc
-Revises: d7db8fd35b41
-Create Date: 2018-08-24 11:28:30.894412
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '14cd800904bc'
-down_revision = 'd7db8fd35b41'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('task_order', sa.Column('attachment_id', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'task_order', 'attachments', ['attachment_id'], ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'task_order', type_='foreignkey')
-    op.drop_column('task_order', 'attachment_id')
-    # ### end Alembic commands ###
