--- alembic/versions/1f57f784ed5b_update_user_from_authnid.py
@@ -1,38 +1,0 @@
-"""update_user_from_authnid
-
-Revision ID: 1f57f784ed5b
-Revises: 55ba973d08b9
-Create Date: 2018-07-30 16:53:05.945005
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '1f57f784ed5b'
-down_revision = '55ba973d08b9'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('dod_id', sa.String(), nullable=True))
-    op.add_column('users', sa.Column('email', sa.String(), nullable=True))
-    op.add_column('users', sa.Column('first_name', sa.String(), nullable=True))
-    op.add_column('users', sa.Column('last_name', sa.String(), nullable=True))
-    op.create_unique_constraint('users_dod_id_unique', 'users', ['dod_id'])
-    op.create_unique_constraint('users_email_unqiue', 'users', ['email'])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('users_dod_id_unique', 'users', type_='unique')
-    op.drop_constraint('users_email_unqiue', 'users', type_='unique')
-    op.drop_column('users', 'last_name')
-    op.drop_column('users', 'first_name')
-    op.drop_column('users', 'email')
-    op.drop_column('users', 'dod_id')
-    # ### end Alembic commands ###
