--- alembic/versions/4ede1e3e50d1_add_authz_models.py
@@ -1,62 +1,0 @@
-"""add_authz_models
-
-Revision ID: 4ede1e3e50d1
-Revises: b5b17d465166
-Create Date: 2018-07-30 11:34:12.016857
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = '4ede1e3e50d1'
-down_revision = 'b5b17d465166'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('roles',
-    sa.Column('id', postgresql.UUID(as_uuid=True), server_default=sa.text('uuid_generate_v4()'), nullable=False),
-    sa.Column('name', sa.String(), nullable=True),
-    sa.Column('description', sa.String(), nullable=True),
-    sa.Column('permissions', postgresql.ARRAY(sa.String()), server_default='{}', nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_roles_name'), 'roles', ['name'], unique=True)
-    op.create_index(op.f('ix_roles_permissions'), 'roles', ['permissions'], unique=False)
-    op.create_table('users',
-    sa.Column('id', postgresql.UUID(as_uuid=True), server_default=sa.text('uuid_generate_v4()'), nullable=False),
-    sa.Column('username', sa.String(), nullable=True),
-    sa.Column('atat_role_id', postgresql.UUID(as_uuid=True), nullable=True),
-    sa.ForeignKeyConstraint(['atat_role_id'], ['roles.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('workspace_role',
-    sa.Column('id', postgresql.UUID(as_uuid=True), server_default=sa.text('uuid_generate_v4()'), nullable=False),
-    sa.Column('workspace_id', postgresql.UUID(as_uuid=True), nullable=True),
-    sa.Column('role_id', postgresql.UUID(as_uuid=True), nullable=True),
-    sa.Column('user_id', postgresql.UUID(as_uuid=True), nullable=True),
-    sa.ForeignKeyConstraint(['role_id'], ['roles.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_workspace_role_user_id'), 'workspace_role', ['user_id'], unique=False)
-    op.create_index(op.f('ix_workspace_role_workspace_id'), 'workspace_role', ['workspace_id'], unique=False)
-    op.create_index('workspace_role_user_workspace', 'workspace_role', ['user_id', 'workspace_id'], unique=True)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('workspace_role_user_workspace', table_name='workspace_role')
-    op.drop_index(op.f('ix_workspace_role_workspace_id'), table_name='workspace_role')
-    op.drop_index(op.f('ix_workspace_role_user_id'), table_name='workspace_role')
-    op.drop_table('workspace_role')
-    op.drop_table('users')
-    op.drop_index(op.f('ix_roles_permissions'), table_name='roles')
-    op.drop_index(op.f('ix_roles_name'), table_name='roles')
-    op.drop_table('roles')
-    # ### end Alembic commands ###
