--- alembic/versions/4f4defb7b440_add_expiration_date_to_task_order.py
@@ -1,28 +1,0 @@
-"""add expiration date to task order
-
-Revision ID: 4f4defb7b440
-Revises: 4c425f17bfe8
-Create Date: 2018-09-17 15:22:33.240310
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '4f4defb7b440'
-down_revision = '4c425f17bfe8'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('task_order', sa.Column('expiration_date', sa.Date(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('task_order', 'expiration_date')
-    # ### end Alembic commands ###
