--- alembic/versions/55ba973d08b9_add_fundz_models.py
@@ -1,38 +1,0 @@
-"""add_fundz_models
-
-Revision ID: 55ba973d08b9
-Revises: 4ea5917e7781
-Create Date: 2018-07-30 14:43:34.099799
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '55ba973d08b9'
-down_revision = '4ea5917e7781'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('pe_number',
-    sa.Column('number', sa.String(), nullable=False),
-    sa.Column('description', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('number')
-    )
-    op.create_table('task_order',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('number', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('task_order')
-    op.drop_table('pe_number')
-    # ### end Alembic commands ###
