--- alembic/versions/777ded5c57a0_bigint_for_request_review_id.py
@@ -1,36 +1,0 @@
-"""bigint for request review id
-
-Revision ID: 777ded5c57a0
-Revises: 7bdb2055d7c7
-Create Date: 2018-09-10 13:24:36.328610
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '777ded5c57a0'
-down_revision = '7bdb2055d7c7'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('request_reviews', 'id',
-               existing_type=sa.Integer(),
-               type_=sa.BigInteger(),
-               nullable=False)
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('request_reviews', 'id',
-               existing_type=sa.BigInteger(),
-               type_=sa.Integer(),
-               nullable=False)
-    pass
-    # ### end Alembic commands ###
