--- alembic/versions/a496111da617_add_name_to_request_revision.py
@@ -1,28 +1,0 @@
-"""add name to request revision
-
-Revision ID: a496111da617
-Revises: 4f4defb7b440
-Create Date: 2018-09-18 16:45:59.297289
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'a496111da617'
-down_revision = '4f4defb7b440'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('request_revisions', sa.Column('name', sa.String(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('request_revisions', 'name')
-    # ### end Alembic commands ###
