--- alembic/versions/a903ebe91ad5_add_sequence_to_request_revision.py
@@ -1,28 +1,0 @@
-"""add sequence to request revision
-
-Revision ID: a903ebe91ad5
-Revises: 04fe150da553
-Create Date: 2018-08-30 13:45:35.561657
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = 'a903ebe91ad5'
-down_revision = '04fe150da553'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    db = op.get_bind()
-    op.add_column('request_revisions', sa.Column('sequence', sa.BigInteger(), nullable=False))
-    db.execute("CREATE SEQUENCE request_revisions_sequence_seq OWNED BY request_revisions.sequence;")
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    op.drop_column('request_revisions', 'sequence')
