--- alembic/versions/f549c7cee17c_remove_workspaces_task_order_association.py
@@ -1,30 +1,0 @@
-"""remove workspaces task order association
-
-Revision ID: f549c7cee17c
-Revises: 4be312655ceb
-Create Date: 2018-08-16 16:42:48.581510
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'f549c7cee17c'
-down_revision = '4be312655ceb'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('workspaces_task_order_id_fkey', 'workspaces', type_='foreignkey')
-    op.drop_column('workspaces', 'task_order_id')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('workspaces', sa.Column('task_order_id', sa.INTEGER(), autoincrement=False))
-    op.create_foreign_key('workspaces_task_order_id_fkey', 'workspaces', 'task_order', ['task_order_id'], ['id'])
-    # ### end Alembic commands ###
